Objects were invisible when viewed from certain angles.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@898 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/wall.h
@@ -141,9 +141,9 @@ typedef struct v19_door {
 //End old tWall structures
 
 typedef struct tWall {
-	short   nSegment, nSide;		// Seg & tSide for this tWall
+	int	  nSegment, nSide;		// Seg & tSide for this tWall
 	fix     hps;						// "Hit points" of the tWall.
-	short   nLinkedWall;				// number of linked tWall
+	int     nLinkedWall;				// number of linked tWall
 	ubyte   nType;						// What kind of special tWall.
 	ubyte   flags;						// Flags for the tWall.
 	ubyte   state;						// Opening, closing, etc.
